+++ previous-versions/commerce/src/sdk/models/operations/listcustomers.ts
@@ -59,4 +59,6 @@ export class ListCustomersResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListCustomersResponse | null>;
 }