+++ previous-versions/commerce/src/sdk/models/operations/listdisputes.ts
@@ -59,4 +59,6 @@ export class ListDisputesResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListDisputesResponse | null>;
 }