+++ previous-versions/commerce/src/sdk/models/operations/listorders.ts
@@ -59,4 +59,6 @@ export class ListOrdersResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListOrdersResponse | null>;
 }