+++ previous-versions/commerce/src/sdk/models/operations/listpaymentmethods.ts
@@ -59,4 +59,6 @@ export class ListPaymentMethodsResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListPaymentMethodsResponse | null>;
 }