+++ previous-versions/commerce/src/sdk/models/operations/listproductcategories.ts
@@ -59,4 +59,6 @@ export class ListProductCategoriesResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListProductCategoriesResponse | null>;
 }