+++ previous-versions/commerce/src/sdk/models/operations/listproducts.ts
@@ -59,4 +59,6 @@ export class ListProductsResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    next: () => Promise<ListProductsResponse | null>;
 }