+++ previous-versions/commerce/src/sdk/models/operations/listtaxcomponents.ts
@@ -59,4 +59,6 @@ export class ListTaxComponentsResponse extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     taxComponents?: shared.TaxComponents;
+
+    next: () => Promise<ListTaxComponentsResponse | null>;
 }