+++ previous-versions/commerce/src/sdk/sdk.ts
@@ -61,8 +61,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.34.0";
-    genVersion = "2.107.3";
+    sdkVersion = "0.34.1";
+    genVersion = "2.108.3";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);